Rearrange root routes
