+++ package.json
@@ -5,7 +5,8 @@
   "private": false,
   "license": "MIT",
   "dependencies": {
-    "@prisma/client": "^2.16.1",
+    "@prisma/client": "^2.17.0",
+    "@sentry/node": "^6.1.0",
     "@slack/web-api": "^6.0.0",
     "@types/bcryptjs": "^2.4.2",
     "@types/cors": "^2.8.10",
@@ -37,7 +38,7 @@
     "nanoid": "^3.1.20",
     "nodemailer": "^6.4.18",
     "pdfkit": "^0.11.0",
-    "prisma": "^2.16.1",
+    "prisma": "^2.17.0",
     "qrcode": "^1.4.4",
     "split": "^1.0.1",
     "swagger-cli": "^4.0.4",
@@ -53,8 +54,8 @@
     "@types/sinon": "^9.0.10",
     "@types/smtp-server": "^3.5.6",
     "@types/supertest": "^2.0.10",
-    "@typescript-eslint/eslint-plugin": "^4.15.0",
-    "@typescript-eslint/parser": "^4.15.0",
+    "@typescript-eslint/eslint-plugin": "^4.15.1",
+    "@typescript-eslint/parser": "^4.15.1",
     "chai": "^4.3.0",
     "chai-string": "^1.5.0",
     "eslint": "^7.20.0",
@@ -63,7 +64,7 @@
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-security": "^1.4.0",
-    "eslint-plugin-unicorn": "^28.0.0",
+    "eslint-plugin-unicorn": "^28.0.1",
     "faker": "^5.4.0",
     "mocha": "^8.3.0",
     "nyc": "^15.1.0",
