+++ src/controllers/index.ts
@@ -2,9 +2,7 @@ import { Router } from 'express';
 import swagger from 'swagger-ui-express';
 
 import swaggerDocument from '../../openapi.json';
-import status from './status';
-import contact from './contact';
-import settings from './settings';
+import root from './root';
 import users from './users';
 import tournaments from './tournaments';
 import items from './items';
@@ -14,14 +12,8 @@ import callbacks from './callbacks';
 
 const router = Router();
 
-// To match only with root
-router.get('//', status);
-
-// Settings route
-router.get('/settings', settings);
-
-// Contact routes
-router.post('/contact', contact);
+// Root routes
+router.use(root);
 
 // Auth routes
 router.use('/auth', auth);
