+++ src/controllers/root/contact.ts
@@ -1,10 +1,10 @@
 import { NextFunction, Request, Response } from 'express';
 import Joi from 'joi';
 
-import { validateBody } from '../middlewares/validation';
-import { noContent } from '../utils/responses';
-import { sendSlackContact } from '../services/slack';
-import * as validators from '../utils/validators';
+import { validateBody } from '../../middlewares/validation';
+import { noContent } from '../../utils/responses';
+import { sendSlackContact } from '../../services/slack';
+import * as validators from '../../utils/validators';
 
 export default [
   // Middleware
