+++ src/controllers/root/settings.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
-import { fetchSettings } from '../operations/settings';
+import { fetchSettings } from '../../operations/settings';
 
-import { Setting } from '../types';
-import { success } from '../utils/responses';
+import { Setting } from '../../types';
+import { success } from '../../utils/responses';
 
 export default [
   async (request: Request, response: Response, next: NextFunction) => {
