+++ src/server.ts
@@ -6,6 +6,7 @@ try {
   // Listen the API on port 3000 (default)
   app.listen(env.api.port, () => {
     logger.info(`Node environment: ${process.env.NODE_ENV}`);
+    logger.info(`Environment: ${env.environment}`);
     logger.info(`Listening on port ${env.api.port}`);
   });
 } catch (error) {
