+++ src/utils/env.ts
@@ -32,6 +32,8 @@ const env = {
   development: process.env.NODE_ENV === 'development',
   production: process.env.NODE_ENV === 'production',
   test: process.env.NODE_ENV === 'test',
+  // Defines the environment used by sentry
+  environment: (loadEnv('ENVIRONMENT') || notInProduction('development')) as 'development' | 'staging' | 'production',
   api: {
     port: loadIntEnv('API_PORT') || 3000,
     prefix: loadEnv('API_PREFIX') || '/',
@@ -103,6 +105,7 @@ const env = {
   },
   log: {
     level: loadEnv('LOG_LEVEL') || 'silly',
+    sentryDsn: loadEnv('LOG_SENTRY_DSN'),
   },
 };
